more fixes for cyclical
