+++ lib/apply-conditions.js
@@ -1,9 +1,7 @@
 "use strict"
 
 const assignLayerNames = require("./assign-layer-names")
-const joinLayer = require("./join-layer")
-const joinMedia = require("./join-media")
-const joinSupports = require("./join-supports")
+const joinConditions = require("./join-conditions")
 
 module.exports = function applyConditions(bundle, options, state, atRule) {
   bundle.forEach(stmt => {
@@ -20,42 +18,7 @@ module.exports = function applyConditions(bundle, options, state, atRule) {
     }
 
     if (stmt.type === "import") {
-      const parts = [stmt.fullUri]
-
-      let media = []
-      let supports = []
-      let layer = []
-
-      for (const condition of stmt.conditions) {
-        media = joinMedia(media, condition.media)
-        supports = joinSupports(supports, condition.supports)
-        layer = joinLayer(layer, condition.layer)
-      }
-
-      if (layer.length) {
-        const layerName = layer.join(".")
-
-        let layerParams = "layer"
-        if (layerName) {
-          layerParams = `layer(${layerName})`
-        }
-
-        parts.push(layerParams)
-      }
-
-      if (supports.length) {
-        if (supports.length === 1) {
-          parts.push(`supports(${supports[0]})`)
-        } else {
-          parts.push(`supports(${supports.map(x => `(${x})`).join(" and ")})`)
-        }
-      }
-
-      if (media.length) {
-        parts.push(media.join(", "))
-      }
-
-      stmt.node.params = parts.join(" ")
+      stmt.node.params = `${stmt.fullUri} ${joinConditions(stmt.conditions)}`
 
       return
     }
