+++ lib/join-media.js
@@ -26,5 +26,5 @@ module.exports = function joinMedia(parentMedia, childMedia) {
     })
   })
 
-  return media
+  return Array.from(new Set(media))
 }
