+++ lib/parse-styles.js
@@ -7,6 +7,7 @@ const dataURL = require("./data-url")
 const parseStatements = require("./parse-statements")
 const processContent = require("./process-content")
 const resolveId = require("./resolve-id")
+const joinConditions = require("./join-conditions")
 
 async function parseStyles(
   result,
@@ -158,7 +159,9 @@ async function loadImportContent(
       state.importedFiles[filename] = {}
     }
     state.importedFiles[filename][stmtDuplicateCheckKey] = true
-  } else if (from.includes(filename)) {
+  }
+
+  if (from.includes(filename)) {
     return
   }
 
@@ -209,19 +212,13 @@ async function loadImportContent(
     options,
     state,
     conditions,
-    [...stmt.from, filename],
+    [...from, filename],
     postcss
   )
 }
 
 function createStmtDuplicateCheckKey(stmt) {
-  return JSON.stringify(
-    stmt.conditions.map(x => [
-      x.media.join("|"),
-      x.layer.join("|"),
-      x.supports.join("|"),
-    ])
-  )
+  return joinConditions(stmt.conditions)
 }
 
 function isProcessableURL(uri) {
